Working things! :D

+++ _Controls/Include/SettingReader.h
@@ -15,8 +15,8 @@ public:
 	static std::string ReadControlMode();
 	static void CheckControlMode(Control*);
 	static std::map<std::string, bool> CheckChangedSettings();
-	static void WriteSensorValues(std::map < std::string, int>);
-	static void LogSensorValues(std::map <std::string, int>);
+	static void WriteSensorValues(std::map < std::string, double>);
+	static void LogSensorValues(std::map <std::string, double>);
 	static void CreateAFile(std::string);
 };
